+++ xclim/indices/run_length.py
@@ -1,5 +1,6 @@
 # noqa: D205,D400
 """
+# TODO: Remove default values from
 Run length algorithms submodule
 ===============================
 
@@ -30,6 +31,7 @@ def use_ufunc(
     ufunc_1dim: bool | str,
     da: xr.DataArray,
     dim: str = "time",
+    freq: str | None = None,
     index: str = "first",
 ) -> bool:
     """Return whether the ufunc version of run length algorithms should be used with this DataArray or not.
@@ -56,13 +58,65 @@ def use_ufunc(
         If ufunc_1dim is "auto", returns True if the array is on dask or too large.
         Otherwise, returns ufunc_1dim.
     """
+    if ufunc_1dim is True and freq is not None:
+        raise ValueError(
+            "Resampling after run length operations is not implemented for 1d method"
+        )
+
     if ufunc_1dim == "from_context":
         ufunc_1dim = OPTIONS[RUN_LENGTH_UFUNC]
 
     if ufunc_1dim == "auto":
         ufunc_1dim = not uses_dask(da) and (da.size // da[dim].size) < npts_opt
+    # If resampling after run length is set up for the computation, the 1d method is not implemented
+    # Unless ufunc_1dim is specifically set to False (in which case we flag an error above),
+    # we simply forbid this possibility.
+    return (index == "first") and (ufunc_1dim) and (freq is None)
 
-    return index == "first" and ufunc_1dim
+
+def resample_and_rl(
+    da: xr.DataArray,
+    resample_before_rl: bool,
+    compute,
+    *args,
+    freq: str,
+    dim: str = "time",
+    **kwargs,
+) -> xr.DataArray | xr.Dataset:
+    """Wrap run length algorithms to control if resampling occurs before or after the algorithms.
+
+    If resample_before_rl is 'from_context', the parameter is read from xclim's global (or context) options.
+
+    Parameters
+    ----------
+    da: xr.DataArray
+      N-dimensional array (boolean).
+    resample_before_rl : bool
+      Determines whether if input arrays of runs `da` should be separated in period before
+      or after the run length algorithms are applied
+    compute
+      Run length function to apply
+    args
+      Positional arguments needed in `compute`
+    dim: str
+      The dimension along which to find runs.
+    freq : str
+      Resampling frequency.
+    kwargs
+      Keyword arguments needed in `compute`
+
+    Returns
+    -------
+    xr.DataArray
+      Output of compute resampled according to frequency {freq}.
+    """
+    if resample_before_rl:
+        out = da.resample({dim: freq}).map(
+            compute, args=args, freq=None, dim=dim, **kwargs
+        )
+    else:
+        out = compute(da, *args, dim=dim, freq=freq, **kwargs)
+    return out
 
 
 def _cumsum_reset_on_zero(
@@ -139,9 +193,10 @@ def rle(
 
 def rle_statistics(
     da: xr.DataArray,
-    reducer: str = "max",
-    window: int = 1,
+    reducer: str,
+    window: int,
     dim: str = "time",
+    freq: str | None = None,
     ufunc_1dim: str | bool = "from_context",
     index: str = "first",
 ) -> xr.DataArray:
@@ -156,7 +211,9 @@ def rle_statistics(
     window : int
         Minimal length of consecutive runs to be included in the statistics.
     dim : str
-        Dimension along which to calculate consecutive run; Default: 'time'.
+      Dimension along which to calculate consecutive run; Default: 'time'.
+    freq : str
+      Resampling frequency.
     ufunc_1dim : Union[str, bool]
         Use the 1d 'ufunc' version of this function : default (auto) will attempt to select optimal
         usage based on number of data points.  Using 1D_ufunc=True is typically more efficient
@@ -172,15 +229,21 @@ def rle_statistics(
         Length of runs of True values along dimension, according to the reducing function (float)
         If there are no runs (but the data is valid), returns 0.
     """
-    ufunc_1dim = use_ufunc(ufunc_1dim, da, dim=dim, index=index)
-    if window == 1 and reducer == "sum":
-        rl_stat = da.sum("time")
-    elif ufunc_1dim:
+    ufunc_1dim = use_ufunc(ufunc_1dim, da, dim=dim, index=index, freq=freq)
+    if ufunc_1dim:
         rl_stat = statistics_run_ufunc(da, reducer, window, dim)
     else:
         d = rle(da, dim=dim, index=index)
-        rl_stat = getattr(d.where(d >= window), reducer)(dim=dim)
-        rl_stat = xr.where((d.isnull() | (d < window)).all(dim=dim), 0, rl_stat)
+
+        def get_rl_stat(d):
+            rl_stat = getattr(d.where(d >= window), reducer)(dim=dim)
+            rl_stat = xr.where((d.isnull() | (d < window)).all(dim=dim), 0, rl_stat)
+            return rl_stat
+
+        if freq is None:
+            rl_stat = get_rl_stat(d)
+        else:
+            rl_stat = d.resample({dim: freq}).map(get_rl_stat)
 
     return rl_stat
 
@@ -188,6 +251,7 @@ def rle_statistics(
 def longest_run(
     da: xr.DataArray,
     dim: str = "time",
+    freq: str | None = None,
     ufunc_1dim: str | bool = "from_context",
     index: str = "first",
 ) -> xr.DataArray:
@@ -198,7 +262,9 @@ def longest_run(
     da : xr.DataArray
         N-dimensional array (boolean).
     dim : str
-        Dimension along which to calculate consecutive run; Default: 'time'.
+      Dimension along which to calculate consecutive run; Default: 'time'.
+    freq : str
+      Resampling frequency.
     ufunc_1dim : Union[str, bool]
         Use the 1d 'ufunc' version of this function : default (auto) will attempt to select optimal
         usage based on number of data points.  Using 1D_ufunc=True is typically more efficient
@@ -214,7 +280,13 @@ def longest_run(
         Length of the longest run of True values along dimension (int).
     """
     return rle_statistics(
-        da, reducer="max", dim=dim, ufunc_1dim=ufunc_1dim, index=index
+        da,
+        reducer="max",
+        window=1,
+        dim=dim,
+        freq=freq,
+        ufunc_1dim=ufunc_1dim,
+        index=index,
     )
 
 
@@ -222,6 +294,7 @@ def windowed_run_events(
     da: xr.DataArray,
     window: int,
     dim: str = "time",
+    freq: str | None = None,
     ufunc_1dim: str | bool = "from_context",
     index: str = "first",
 ) -> xr.DataArray:
@@ -235,7 +308,9 @@ def windowed_run_events(
         Minimum run length.
         When equal to 1, an optimized version of the algorithm is used.
     dim : str
-        Dimension along which to calculate consecutive run (default: 'time').
+      Dimension along which to calculate consecutive run (default: 'time').
+    freq : str
+      Resampling frequency.
     ufunc_1dim : Union[str, bool]
         Use the 1d 'ufunc' version of this function : default (auto) will attempt to select optimal
         usage based on number of data points.  Using 1D_ufunc=True is typically more efficient
@@ -250,16 +325,23 @@ def windowed_run_events(
     xr.DataArray, [int]
         Number of distinct runs of a minimum length (int).
     """
-    ufunc_1dim = use_ufunc(ufunc_1dim, da, dim=dim, index=index)
+    ufunc_1dim = use_ufunc(ufunc_1dim, da, dim=dim, index=index, freq=freq)
 
-    if window == 1:
-        d = da.pad({dim: (0, 1)}, constant_values=False).astype(int).diff(dim)
-        out = (d == -1).sum(dim=dim)
-    elif ufunc_1dim:
+    if ufunc_1dim:
         out = windowed_run_events_ufunc(da, window, dim)
+
     else:
-        d = rle(da, dim=dim, index=index)
-        out = (d >= window).sum(dim=dim)
+        if window == 1:
+            shift = 1 * (index == "first") + -1 * (index == "last")
+            d = xr.where(da.shift({dim: shift}, fill_value=0) == 0, 1, 0)
+            d = d.where(da == 1, 0)
+        else:
+            d = rle(da, dim=dim, index=index)
+            d = xr.where(d >= window, 1, 0)
+        if freq is not None:
+            d = d.resample({dim: freq})
+        out = d.sum(dim=dim)
+
     return out
 
 
@@ -267,6 +349,7 @@ def windowed_run_count(
     da: xr.DataArray,
     window: int,
     dim: str = "time",
+    freq: str | None = None,
     ufunc_1dim: str | bool = "from_context",
     index: str = "first",
 ) -> xr.DataArray:
@@ -280,7 +363,9 @@ def windowed_run_count(
         Minimum run length.
         When equal to 1, an optimized version of the algorithm is used.
     dim : str
-        Dimension along which to calculate consecutive run (default: 'time').
+      Dimension along which to calculate consecutive run (default: 'time').
+    freq : str
+      Resampling frequency.
     ufunc_1dim : Union[str, bool]
         Use the 1d 'ufunc' version of this function : default (auto) will attempt to select optimal
         usage based on number of data points. Using 1D_ufunc=True is typically more efficient
@@ -297,13 +382,19 @@ def windowed_run_count(
     """
     ufunc_1dim = use_ufunc(ufunc_1dim, da, dim=dim, index=index)
 
-    if window == 1:
-        out = da.sum(dim=dim)
-    elif ufunc_1dim:
+    if ufunc_1dim:
         out = windowed_run_count_ufunc(da, window, dim)
+
+    elif window == 1 and freq is None:
+        out = da.sum(dim=dim)
+
     else:
         d = rle(da, dim=dim, index=index)
-        out = d.where(d >= window, 0).sum(dim=dim)
+        d = d.where(d >= window, 0)
+        if freq is not None:
+            d = d.resample({dim: freq})
+        out = d.sum(dim=dim)
+
     return out
 
 
@@ -311,6 +402,7 @@ def first_run(
     da: xr.DataArray,
     window: int,
     dim: str = "time",
+    freq: str | None = None,
     coord: str | bool | None = False,
     ufunc_1dim: str | bool = "from_context",
 ) -> xr.DataArray:
@@ -324,7 +416,9 @@ def first_run(
         Minimum duration of consecutive run to accumulate values.
         When equal to 1, an optimized version of the algorithm is used.
     dim : str
-        Dimension along which to calculate consecutive run (default: 'time').
+      Dimension along which to calculate consecutive run (default: 'time').
+    freq : str
+      Resampling frequency.
     coord : Optional[str]
         If not False, the function returns values along `dim` instead of indexes.
         If `dim` has a datetime dtype, `coord` can also be a str of the name of the
@@ -341,31 +435,48 @@ def first_run(
         Index (or coordinate if `coord` is not False) of first item in first valid run.
         Returns np.nan if there are no valid runs.
     """
-    ufunc_1dim = use_ufunc(ufunc_1dim, da, dim=dim)
+    # transforms indexes to coordinates if needed, and drops obsolete dim
+    def coord_transform(out, da):
+        if coord:
+            crd = da[dim]
+            if isinstance(coord, str):
+                crd = getattr(crd.dt, coord)
 
-    da = da.fillna(0)  # We expect a boolean array, but there could be NaNs nonetheless
-    if ufunc_1dim:
-        out = first_run_ufunc(x=da, window=window, dim=dim)
-    else:
-        if window == 1:
-            d = da
-        else:
-            d = rle(da, dim=dim, index="first")
-            d = xr.where(d >= window, 1, -1)
+            out = lazy_indexing(crd, out)
 
+        if dim in out.coords:
+            out = out.drop_vars(dim)
+        return out
+
+    # general method to get indices (or coords) of first run
+    def get_out(d):
         dmax_ind = d.argmax(dim=dim)
         # If `d` has no runs, dmax_ind will be 0: We must replace this with NaN
         out = dmax_ind.where(dmax_ind != d.argmin(dim=dim))
+        out = coord_transform(out, d)
+        return out
 
-    if coord:
-        crd = da[dim]
-        if isinstance(coord, str):
-            crd = getattr(crd.dt, coord)
+    ufunc_1dim = use_ufunc(ufunc_1dim, da, dim=dim, freq=freq)
 
-        out = lazy_indexing(crd, out)
+    da = da.fillna(0)  # We expect a boolean array, but there could be NaNs nonetheless
+    if window == 1:
+        if freq is not None:
+            out = da.resample({dim: freq}).map(get_out)
+        else:
+            out = xr.where(da.any(dim=dim), da.argmax(dim=dim), np.NaN)
+            out = coord_transform(out, da)
 
-    if dim in out.coords:
-        out = out.drop_vars(dim)
+    elif ufunc_1dim:
+        out = first_run_ufunc(x=da, window=window, dim=dim)
+        out = coord_transform(out, da)
+
+    else:
+        d = rle(da, dim=dim, index="first")
+        d = xr.where(d >= window, 1, 0)
+        if freq is not None:
+            out = d.resample({dim: freq}).map(get_out)
+        else:
+            out = get_out(d)
 
     return out
 
@@ -374,6 +485,7 @@ def last_run(
     da: xr.DataArray,
     window: int,
     dim: str = "time",
+    freq: str | None = None,
     coord: str | bool | None = False,
     ufunc_1dim: str | bool = "from_context",
 ) -> xr.DataArray:
@@ -387,7 +499,9 @@ def last_run(
         Minimum duration of consecutive run to accumulate values.
         When equal to 1, an optimized version of the algorithm is used.
     dim : str
-        Dimension along which to calculate consecutive run (default: 'time').
+      Dimension along which to calculate consecutive run (default: 'time').
+    freq : str
+      Resampling frequency.
     coord : Optional[str]
         If not False, the function returns values along `dim` instead of indexes.
         If `dim` has a datetime dtype, `coord` can also be a str of the name of the
@@ -406,7 +520,12 @@ def last_run(
     """
     reversed_da = da.sortby(dim, ascending=False)
     out = first_run(
-        reversed_da, window=window, dim=dim, coord=coord, ufunc_1dim=ufunc_1dim
+        reversed_da,
+        window=window,
+        dim=dim,
+        freq=freq,
+        coord=coord,
+        ufunc_1dim=ufunc_1dim,
     )
     if not coord:
         return reversed_da[dim].size - out - 1
@@ -478,7 +597,9 @@ def run_bounds(mask: xr.DataArray, dim: str = "time", coord: bool | str = True):
     return xr.concat((starts, ends), "bounds")
 
 
-def keep_longest_run(da: xr.DataArray, dim: str = "time") -> xr.DataArray:
+def keep_longest_run(
+    da: xr.DataArray, dim: str = "time", freq: str | None = None
+) -> xr.DataArray:
     """Keep the longest run along a dimension.
 
     Parameters
@@ -486,7 +607,9 @@ def keep_longest_run(da: xr.DataArray, dim: str = "time") -> xr.DataArray:
     da : xr.DataArray
         Boolean array.
     dim : str
-        Dimension along which to check for the longest run.
+      Dimension along which to check for the longest run.
+    freq : str
+      Resampling frequency.
 
     Returns
     -------
@@ -495,14 +618,23 @@ def keep_longest_run(da: xr.DataArray, dim: str = "time") -> xr.DataArray:
     """
     # Get run lengths
     rls = rle(da, dim)
-    out = xr.where(
-        # Construct an integer array and find the max
-        rls[dim].copy(data=np.arange(rls[dim].size)) == rls.argmax(dim),
-        rls + 1,  # Add one to the First longest run
-        rls,
-    )
-    out = out.ffill(dim) == out.max(dim)
-    return da.copy(data=out.transpose(*da.dims).data)  # Keep everything the same
+
+    def get_out(rls):
+        out = xr.where(
+            # Construct an integer array and find the max
+            rls[dim].copy(data=np.arange(rls[dim].size)) == rls.argmax(dim),
+            rls + 1,  # Add one to the First longest run
+            rls,
+        )
+        out = out.ffill(dim) == out.max(dim)
+        return out
+
+    if freq is not None:
+        out = rls.resample({dim: freq}).map(get_out)
+    else:
+        out = get_out(rls)
+
+    return da.copy(data=out.transpose(*da.dims).data)
 
 
 def season(
@@ -873,7 +1005,7 @@ def first_run_1d(arr: Sequence[int | float], window: int) -> int | np.nan:
     return ind
 
 
-def statistics_run_1d(arr: Sequence[bool], reducer: str, window: int = 1) -> int:
+def statistics_run_1d(arr: Sequence[bool], reducer: str, window: int) -> int:
     """Return statistics on lengths of run of identical values.
 
     Parameters
@@ -1000,7 +1132,7 @@ def windowed_run_events_ufunc(
 def statistics_run_ufunc(
     x: xr.DataArray | Sequence[bool],
     reducer: str,
-    window: int = 1,
+    window: int,
     dim: str = "time",
 ) -> xr.DataArray:
     """Dask-parallel version of statistics_run_1d, ie: the {reducer} number of consecutive true values in array.